requirements.txt
google_api_python_client==2.119.0
protobuf==5.27.1
streamlit==1.35.0
